+++ mappers/Mapper019.h
@@ -23,24 +23,24 @@ public:
 public:
 	std::string name() const override;
 
-	uint8_t read_4(uint16_t address) override;
-	uint8_t read_5(uint16_t address) override;
-	uint8_t read_6(uint16_t address) override;
-	uint8_t read_7(uint16_t address) override;
+	uint8_t read_4(uint32_t address) override;
+	uint8_t read_5(uint32_t address) override;
+	uint8_t read_6(uint32_t address) override;
+	uint8_t read_7(uint32_t address) override;
 
 public:
-	void write_4(uint16_t address, uint8_t value) override;
-	void write_5(uint16_t address, uint8_t value) override;
-	void write_6(uint16_t address, uint8_t value) override;
-	void write_7(uint16_t address, uint8_t value) override;
-	void write_8(uint16_t address, uint8_t value) override;
-	void write_9(uint16_t address, uint8_t value) override;
-	void write_a(uint16_t address, uint8_t value) override;
-	void write_b(uint16_t address, uint8_t value) override;
-	void write_c(uint16_t address, uint8_t value) override;
-	void write_d(uint16_t address, uint8_t value) override;
-	void write_e(uint16_t address, uint8_t value) override;
-	void write_f(uint16_t address, uint8_t value) override;
+	void write_4(uint32_t address, uint8_t value) override;
+	void write_5(uint32_t address, uint8_t value) override;
+	void write_6(uint32_t address, uint8_t value) override;
+	void write_7(uint32_t address, uint8_t value) override;
+	void write_8(uint32_t address, uint8_t value) override;
+	void write_9(uint32_t address, uint8_t value) override;
+	void write_a(uint32_t address, uint8_t value) override;
+	void write_b(uint32_t address, uint8_t value) override;
+	void write_c(uint32_t address, uint8_t value) override;
+	void write_d(uint32_t address, uint8_t value) override;
+	void write_e(uint32_t address, uint8_t value) override;
+	void write_f(uint32_t address, uint8_t value) override;
 
 public:
 	void cpu_sync() override;
