+++ mappers/Mapper025.h
@@ -12,17 +12,17 @@ public:
 	std::string name() const override;
 
 public:
-	uint8_t read_6(uint16_t address) override;
+	uint8_t read_6(uint32_t address) override;
 
 public:
-	void write_6(uint16_t address, uint8_t value) override;
-	void write_8(uint16_t address, uint8_t value) override;
-	void write_9(uint16_t address, uint8_t value) override;
-	void write_a(uint16_t address, uint8_t value) override;
-	void write_b(uint16_t address, uint8_t value) override;
-	void write_c(uint16_t address, uint8_t value) override;
-	void write_d(uint16_t address, uint8_t value) override;
-	void write_e(uint16_t address, uint8_t value) override;
+	void write_6(uint32_t address, uint8_t value) override;
+	void write_8(uint32_t address, uint8_t value) override;
+	void write_9(uint32_t address, uint8_t value) override;
+	void write_a(uint32_t address, uint8_t value) override;
+	void write_b(uint32_t address, uint8_t value) override;
+	void write_c(uint32_t address, uint8_t value) override;
+	void write_d(uint32_t address, uint8_t value) override;
+	void write_e(uint32_t address, uint8_t value) override;
 };
 
 #endif
